+++ packages/media-query-list-parser/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changes to Media Query List Parser
 
+### Unreleased (minor)
+
+- Add `state` to `walk` methods.
+
+This makes it possible pass down information from a parent structure to children.  
+e.g. you can set `entry.state.inInPrintQuery = true` for `print and (min-width: 30cm)`.
+
 ### 2.0.4 (April 10, 2023)
 
 - Updated `@csstools/css-tokenizer` to `2.1.1` (patch)